requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
attrs=22.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_win64_mkl
bleach=6.0.0=pypi_0
brotlipy=0.7.0=py39ha55989b_1005
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h68f70e3_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.2=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.2=py39hb6bd5e6_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.1.55=0
cuda-documentation=12.1.55=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.1.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.1.55=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.1.55=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.1.55=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.1.55=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
dask=2023.3.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.0=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
fonttools=4.38.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h546665d_1
fsspec=2023.3.0=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.0.0=h57928b3_25922
ipykernel=6.21.2=pypi_0
ipython=8.11.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=hcfcfb64_3
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.0.3=pypi_0
jupyter-console=6.6.2=pypi_0
jupyter-core=5.2.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.3.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
kiwisolver=1.4.4=pypi_0
lcms2=2.14=ha5c8aab_1
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.9.0=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=16_win64_mkl
liblapacke=3.9.0=16_win64_mkl
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h19919ed_0
libsqlite=3.40.0=hcfcfb64_0
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.3=hc3477c8_0
libzlib=1.2.13=hcfcfb64_4
locket=1.0.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
nbclassic=0.5.3=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
nest-asyncio=1.5.6=pypi_0
notebook=6.5.3=pypi_0
notebook-shim=0.2.2=pypi_0
nsight-compute=2023.1.0.15=0
numpy=1.24.2=py39h16ffa76_0
openjpeg=2.5.0=ha2aaf27_2
openpyxl=3.1.1=pypi_0
openssl=3.0.8=hcfcfb64_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pandas-plink=2.2.9=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
partd=1.3.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py39hcebd2be_1
pip=23.0.1=pyhd8ed1ab_0
platformdirs=3.1.0=pypi_0
pluggy=1.0.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
psutil=5.9.4=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.14.0=pypi_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pyh0701188_6
pytest=7.2.2=pypi_0
python=3.9.16=h4de0772_0_cpython
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.9=3_cp39
pytorch=1.13.1=py3.9_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pywin32=305=pypi_0
pywinpty=2.0.10=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.0=pypi_0
qtconsole=5.4.0=pypi_0
qtpy=2.3.0=pypi_0
requests=2.28.2=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=67.5.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
stack-data=0.6.2=pypi_0
tbb=2021.8.0=h91493d7_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h8ffe710_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
ucrt=10.0.22621.0=h57928b3_0
uri-template=1.2.0=pypi_0
urllib3=1.26.14=pyhd8ed1ab_0
vc=14.3=hb6edc58_10
vs2015_runtime=14.34.31931=h4c5c07a_10
wcwidth=0.2.6=pypi_0
webcolors=1.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
widgetsnbextension=4.0.5=pypi_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.15.0=pypi_0
xarray=2023.2.0=pypi_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zipp=3.15.0=pypi_0
zstandard=0.20.0=pypi_0
zstd=1.5.2=h12be248_6
